Format, and add editorconfig
I added an .editorconfig file, which is supported by all major editors, and makes everyone have the same formatting settings.

It sounds like krist is a mod, which it is not. If you have a better solution, just let me know.

Additionally I formatted the README according to the spec & markdownlint.